Make the encoding configurable by a function template parameter (traits)
Introduce Traits as the first and mandatory parameter to all encoding
functions to allow configuring the encoding behaviour.  (It needs more
code for *every* function call in the client code, though.)  The traits
provides alphabet, padding and line break position customization.

Move the translation of the 6-bit code the the BASE64 character to the
alphabet static class.

Move the encoded ouptut padding implementation to the padding static class.

Introduce the line break insertion when the maximum line length is exceeded.
Add the currently encoded line length to the encoding state to support
putting line breaks to the encoded output.

Introduce traits template specializations for the most usual cases from the
RFC 4648: normal, url, mime and pem.

Change io::clear_state to a function to make a distinction between it and
the output stream manipulators; a function for emergency scenarios shouldn't
be used the same way as routine outputting manipulators.

Update the tests to check the new interface with encoding traits.

Include the internal BASE64 encoding from cpp-netlib message in the
encoding implementation experiments.

Some code and comment reformatting.