Change the interface from free functions to a static class
Encapsulate the free functions in a static class - converter.
Move the traits template prameters to the converter class;
a new typedef can be created for a specialization of this
class, which can reduce the amount of the client code.

However, it also closes the interface;  free functions could
be easily added for other parameters; the class would have to
be inherited from and the new class name used in the client code.